+++ dzVents/documentation/history.md
@@ -1,3 +1,7 @@
+[2.4.19]
+- Add stringSplit function to domoticz.utils.
+- Add statusText and protocol to HTTPResponse
+
 [2.4.18]
 - Add triggerIFTTT()